Include VS prerelease in vcvars*.bat search
When searching for the Visual Studio installation path the build will fail if the installed version of VS2019 isn't considered a release version. It shouldn't stop the build process, and I'm not sure this is the best approach to solve this problem, but on VS2019 Community v16.9.4 this is the only way I could find to get past this step in the build process without hardcoding the path or receiving it as an argument.